Wordle board is visually working, webcam is installed and detection is working
