+++ client/components/WordleApp.js
@@ -1,6 +1,6 @@
 import React, { Component, useState, createContext } from 'react';
 import WordleBoard from './Wordle/WordleBoard';
-import ASLDetection from './ASLDetection';
+import WordleDetection from './WordleDetection';
 import { boardDefault, rows } from './Wordle/wordleUtilities';
 
 // class WordleApp extends Component {
@@ -9,14 +9,17 @@ import { boardDefault, rows } from './Wordle/wordleUtilities';
 export const WordleAppContext = createContext();
 function WordleApp() {
     const [board, setBoard] = useState(rows);
+    const [translation, setTranslation] = useState(null)
 
     return (
         <div className='wordle-app'>
             <h1>Wordle</h1>
-            <WordleAppContext.Provider value={{ board, setBoard }}>
+            <WordleAppContext.Provider value={{ board, setBoard, translation, setTranslation }}>
                 <div className='game'>
                     <div id='webcam-parent'>
-                        <ASLDetection />
+                        <h2>{translation || 'Hold up a letter in ASL to begin the detection!'}</h2>
+                        {/* <h2>Detecting: {translation}</h2> */}
+                        <WordleDetection />
                     </div>
                     <div id='board-parent'>
                         <WordleBoard />
